requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
llama-cpp-python==0.2.20
pydantic==2.5.1
typing-extensions==4.8.0
python-multipart==0.0.6
starlette==0.27.0
geopy==2.4.1
numpy==2.2.3
pandas==2.2.3
sqlalchemy==2.0.28
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.5  
